Move the TestFramework to the appropriate project
