+++ ProcessDoctor.TestFramework/FakeException.cs
@@ -1,0 +1,4 @@
+namespace ProcessDoctor.TestFramework;
+
+public sealed class FakeException()
+    : Exception("This exception is caused intentionally");
